+++ rsass/src/sass/functions/resolvedargs.rs
@@ -1,5 +1,6 @@
 use super::{CallError, CheckedArg, Name};
 use crate::css::Value;
+use crate::value::ListSeparator;
 use crate::ScopeRef;
 
 /// The arguments to a builtin function.
@@ -23,6 +24,34 @@ impl ResolvedArgs {
         <T as TryFrom<Value>>::Error: ToString,
     {
         self.get_map(name, |v| T::try_from(v).map_err(|e| e.to_string()))
+    }
+
+    /// Get a checked var-args parameter as a Vec of a given type.
+    pub fn get_va<T>(&self, name: Name) -> Result<Vec<T>, CallError>
+    where
+        T: TryFrom<Value>,
+        <T as TryFrom<Value>>::Error: ToString,
+    {
+        fn inner<T>(value: Value) -> Result<Vec<T>, String>
+        where
+            T: TryFrom<Value>,
+            <T as TryFrom<Value>>::Error: ToString,
+        {
+            let check = |value: Value| -> Result<T, String> {
+                T::try_from(value).map_err(|e| e.to_string())
+            };
+            match value {
+                Value::ArgList(args) => {
+                    args.check_no_named().map_err(|e| e.to_string())?;
+                    args.positional.into_iter().map(check).collect()
+                }
+                Value::List(v, Some(ListSeparator::Comma), false) => {
+                    v.into_iter().map(check).collect()
+                }
+                single => Ok(vec![check(single)?]),
+            }
+        }
+        inner(self.scope.get(&name)?).named(name)
     }
 
     /// Get an optional named argument.